@@ -1,6 +1,6 @@
 module github.com/paketo-buildpacks/bundle-install
 
-go 1.18
+go 1.20
 
 require (
 	github.com/Masterminds/semver/v3 v3.2.0
